+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Master
 
+- [BUGFIX] Pressing up/down arrown on selects without a search wasn't default prevented and the page scrolled.
+- [ENHANCEMENT] Setting `theme` to false in the addon config entirely disables the styles.
+
 # 1.0.0-beta.2
 - [ENHANCEMENT] Input placeholders have no opacity now, to be consistent with span-based placeholder.